+++ neo/Prompt/Commands/BuildNRun.py
@@ -1,11 +1,12 @@
 from neo.Prompt.Utils import get_arg
-from neo.Prompt.Commands.LoadSmartContract import GatherLoadedContractParams
-from neo.Prompt.Commands.Invoke import test_deploy_and_invoke
+from neo.Prompt.Commands.LoadSmartContract import GatherLoadedContractParams, generate_deploy_script
+from neo.Prompt.Commands.Invoke import test_deploy_and_invoke, DEFAULT_MIN_FEE
 from neocore.Fixed8 import Fixed8
 from boa.compiler import Compiler
 
 import binascii
 import traceback
+from neo.Core.State.ContractState import ContractPropertyState
 
 
 def LoadAndRun(arguments, wallet):
@@ -32,7 +33,7 @@ def LoadAndRun(arguments, wallet):
         print("Could not load script %s " % e)
 
 
-def BuildAndRun(arguments, wallet, verbose=True):
+def BuildAndRun(arguments, wallet, verbose=True, min_fee=DEFAULT_MIN_FEE):
     path = get_arg(arguments)
 
     contract_script = Compiler.instance().load_and_save(path)
@@ -40,10 +41,10 @@ def BuildAndRun(arguments, wallet, verbose=True):
     newpath = path.replace('.py', '.avm')
     print("Saved output to %s " % newpath)
 
-    return DoRun(contract_script, arguments, wallet, path, verbose)
+    return DoRun(contract_script, arguments, wallet, path, verbose, min_fee=min_fee)
 
 
-def DoRun(contract_script, arguments, wallet, path, verbose=True):
+def DoRun(contract_script, arguments, wallet, path, verbose=True, min_fee=DEFAULT_MIN_FEE):
 
     test = get_arg(arguments, 1)
 
@@ -56,7 +57,7 @@ def DoRun(contract_script, arguments, wallet, path, verbose=True):
 
             script = GatherLoadedContractParams(f_args, contract_script)
 
-            tx, result, total_ops, engine = test_deploy_and_invoke(script, i_args, wallet)
+            tx, result, total_ops, engine = test_deploy_and_invoke(script, i_args, wallet, min_fee)
             i_args.reverse()
 
             if tx is not None and result is not None:
@@ -80,3 +81,15 @@ def DoRun(contract_script, arguments, wallet, path, verbose=True):
             print("please open a wallet to test built contract")
 
     return None, None, None, None
+
+
+def TestBuild(script, invoke_args, wallet, plist='05', ret='05', dynamic=False):
+
+    properties = ContractPropertyState.HasStorage
+
+    if dynamic:
+        properties += ContractPropertyState.HasDynamicInvoke
+
+    script = generate_deploy_script(script, contract_properties=int(properties), parameter_list=plist, return_type=ret)
+
+    return test_deploy_and_invoke(script, invoke_args, wallet)