+++ neo/api/REST/NotificationRestApi.py
@@ -11,15 +11,17 @@ from logzero import logger
 from neo.Core.Blockchain import Blockchain
 from neocore.UInt160 import UInt160
 from neocore.UInt256 import UInt256
+from neo.Settings import settings
+
+
+API_URL_PREFIX = "/v1"
 
 
 class NotificationRestApi(object):
     app = Klein()
-
     notif = None
 
     def __init__(self):
-
         self.notif = NotificationDB.instance()
 
     #
@@ -27,21 +29,28 @@ class NotificationRestApi(object):
     #
     @app.route('/')
     def home(self, request):
+        endpoints_html = """<ul>
+            <li><pre>{apiPrefix}/notifications/block/<height></pre> <em>notifications by block</em></li>
+            <li><pre>{apiPrefix}/notifications/addr/<addr></pre><em>notifications by address</em></li>
+            <li><pre>{apiPrefix}/notifications/tx/<hash></pre><em>notifications by tx</em></li>
+            <li><pre>{apiPrefix}/notifications/contract/<hash></pre><em>notifications by contract</em></li>
+            <li><pre>{apiPrefix}/tokens</pre><em>lists all NEP5 Tokens</em></li>
+            <li><pre>{apiPrefix}/token/<contract_hash></pre><em>list an NEP5 Token</em></li>
+        </ul>
+        """.format(apiPrefix=API_URL_PREFIX)
 
         return """<html>
                     <style>body {padding:20px;max-width:800px;pre { background-color:#eee; }</style>
                     <body>
-                        <h2>endpoints:</h2>
                         <p>
-                            <ul>
-                                <li><pre>/block/{height}</pre> <em>notifications by block</em></li>
-                                <li><pre>/addr/{addr}</pre><em>notifications by address</em></li>
-                                <li><pre>/tx/{hash}</pre><em>notifications by tx</em></li>
-                                <li><pre>/contract/{hash}</pre><em>notifications by contract</em></li>
-                                <li><pre>/tokens</pre><em>lists all NEP5 Tokens</em></li>
-                                <li><pre>/token/{contract_hash}</pre><em>list an NEP5 Token</em></li>                                
-                            </ul>
+                            <h2>REST API for NEO %s</h2>
+                            (see also <a href="https://github.com/CityOfZion/neo-python">neo-python</a>, <a href="https://github.com/CityOfZion/neo-python/blob/development/api-server.py">api-server.py</a>)
                         </p>
+
+                        <hr/>
+
+                        <h2>endpoints:</h2>
+                        <p>%s</p>
                         <div>
                             <hr/>
                             <h3>pagination</h3>
@@ -103,9 +112,9 @@ class NotificationRestApi(object):
 }</pre>
                         </div>
                     </body>
-                </html>"""
+                </html>""" % (settings.net_name, endpoints_html)
 
-    @app.route('/block/<int:block>', methods=['GET'])
+    @app.route('%s/notifications/block/<int:block>' % API_URL_PREFIX, methods=['GET'])
     def get_by_block(self, request, block):
         request.setHeader('Content-Type', 'application/json')
         try:
@@ -115,7 +124,7 @@ class NotificationRestApi(object):
             return self.format_message("Could not get notifications for block %s because %s " % (block, e))
         return self.format_notifications(request, notifications)
 
-    @app.route('/addr/<string:address>', methods=['GET'])
+    @app.route('%s/addr/<string:address>' % API_URL_PREFIX, methods=['GET'])
     def get_by_addr(self, request, address):
         request.setHeader('Content-Type', 'application/json')
         try:
@@ -125,7 +134,7 @@ class NotificationRestApi(object):
             return self.format_message("Could not get notifications for address %s because %s" % (address, e))
         return self.format_notifications(request, notifications)
 
-    @app.route('/tx/<string:tx_hash>', methods=['GET'])
+    @app.route('%s/tx/<string:tx_hash>' % API_URL_PREFIX, methods=['GET'])
     def get_by_tx(self, request, tx_hash):
         request.setHeader('Content-Type', 'application/json')
 
@@ -144,7 +153,7 @@ class NotificationRestApi(object):
 
         return self.format_notifications(request, notifications)
 
-    @app.route('/contract/<string:contract_hash>', methods=['GET'])
+    @app.route('%s/contract/<string:contract_hash>' % API_URL_PREFIX, methods=['GET'])
     def get_by_contract(self, request, contract_hash):
         request.setHeader('Content-Type', 'application/json')
         try:
@@ -155,13 +164,13 @@ class NotificationRestApi(object):
             return self.format_message("Could not get notifications for contract hash %s because %s" % (contract_hash, e))
         return self.format_notifications(request, notifications)
 
-    @app.route('/tokens', methods=['GET'])
+    @app.route('%s/tokens' % API_URL_PREFIX, methods=['GET'])
     def get_tokens(self, request):
         request.setHeader('Content-Type', 'application/json')
         notifications = self.notif.get_tokens()
         return self.format_notifications(request, notifications)
 
-    @app.route('/token/<string:contract_hash>', methods=['GET'])
+    @app.route('%s/token/<string:contract_hash>' % API_URL_PREFIX, methods=['GET'])
     def get_token(self, request, contract_hash):
         request.setHeader('Content-Type', 'application/json')
         try: